+++ cpp/runtimes/libjoynr-runtime/LibJoynrRuntime.cpp
@@ -63,6 +63,8 @@ LibJoynrRuntime::LibJoynrRuntime(std::unique_ptr<Settings> settings)
 
 LibJoynrRuntime::~LibJoynrRuntime()
 {
+    messageSender->registerDispatcher(nullptr);
+
     if (inProcessDispatcher) {
         delete inProcessDispatcher;
         inProcessDispatcher = nullptr;