+++ java/jeeintegration/src/main/java/io/joynr/jeeintegration/api/JoynrMqttClientIdProvider.java
@@ -1,0 +1,44 @@
+package io.joynr.jeeintegration.api;
+
+/*
+ * #%L
+ * %%
+ * Copyright (C) 2011 - 2017 BMW Car IT GmbH
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import javax.inject.Qualifier;
+
+/**
+ * Used to specify an optional mqtt client id provider
+ * <p>
+ * A mqtt client id provider is used to create an id which identifies the mqtt client.
+ * <p>
+ * <b>Note</b> that if the EJB which contains the producer methods implements an interface, then the producer methods
+ * also need to be declared in that interface, otherwise CDI won't recognize the method implementations as producers.
+ */
+@Documented
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ ElementType.METHOD, ElementType.PARAMETER, ElementType.FIELD })
+@Qualifier
+public @interface JoynrMqttClientIdProvider {
+
+}